pyproject.toml
[project]
name = "nautilus-alpaca-adapter"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "alpaca-py>=0.42.0",
    "httpx>=0.28.1",
]

[dependency-groups]
dev = [
    "nautilus-trader>=1.219.0",
    "pyqt6>=6.9.1",
    "pytest>=8.4.1",
    "qtconsole>=5.6.1",
    "ruff>=0.12.3",
]

[tool.ruff]
line-length = 90
target-version = "py312"

[tool.ruff.format]
line-ending = "lf"

[tool.ruff.lint.isort]
case-sensitive = true
lines-after-imports = 2
lines-between-types = 1
required-imports = ["from __future__ import annotations"]
section-order = [
    "future",
    "standard-library",
    "third-party",
    "alpaca",
    "nautilus",
    "first-party",
    "local-folder",
]

[tool.ruff.lint.isort.sections]
# Group all Django imports into a separate section.
"alpaca" = ["alpaca"]
"nautilus" = ["nautilus_trader"]

[tool.pytest.ini_options]
addopts = "-s"
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
